Refactor data validation and auth middleware
